Add files via upload
Write a Java program with four classes in four separate source code files.

The first class is Event. Events should at a minimum have a date, start and end time, location name, location address, event name, number of guests, point of contact, and price. You may add additional attributes.

The second class is BirthdayParty. A BirthdayParty is-a Event, but should also include appropriate attributes and methods for age, cake, candles, and decorations.

The third class is Quinceañera (you may replace the ñ with an English n for purposes of the program). A Quinceañera is a specific type of BirthdayParty. The age should be immutably set to 15. Additional attributes will include number of damas or maids of honor, number of toasts performed, and choice of dance music.

A Quinceañera is-a BirthdayParty.  A BirthdayParty is-a Event.

The fourth class is EventPlanner. EventPlanner will include a runnable main() method.

When run, the program should print out the course, assignment, date, and student name to the screen. Then prompt the user for the number of events to create.  Then create an array of type Event of the length given by the user. For each item in the array, prompt the user for the event type, choosing from the three types. Accept the user’s responses and create an object of that type and assign it to the array.

After all user-supplied objects are created, loop through the array and print out the event planning report showing the state of each event.

All classes should demonstrate inheritance, encapsulation, have appropriate get and set methods, and have a toString() method. Child classes should override toString() and call parent constructors.